+++ java/elisha/project/entity/Item.java
@@ -18,7 +18,7 @@ public class Item {
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
 	@Column(name = "itemId")
-	private Long id;
+	private int id;
 	
 	@Column(nullable = false, unique = true)
 	private String title;
@@ -54,12 +54,12 @@ public class Item {
 		this.quantity = quantity;
 	}
 
-	public Long getId()
+	public int getId()
 	{
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
